pyproject.toml
[tool.poetry]
name = "archive"
version = "0.1.0"
description = "Data collection and analysis tools for CMHLP's Public Archive"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^1.10.7"
httpx = "^0.24.0"
yarl = "^1.9.2"
aiogoogle = "^5.3.0"
isort = "^5.12.0"
google-api-python-client = "^2.90.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
